requirements.txt
pyrogram
TgCrypto
Pyromod
pymongo
dnspython
aiohttp
geezlibs
python-dotenv